Refactored guard import paths in controllers
