+++ src/modules/enviroment/interface/environment-team.controller.ts
@@ -12,9 +12,9 @@ import {
 
 import { IPromiseResponse } from '@/common/response_service/interface/response.interface';
 import { Auth } from '@/modules/auth/application/decorator/auth.decorator';
+import { AdminRoleGuard } from '@/modules/auth/application/guard/admin-role.guard';
+import { AuthTeamGuard } from '@/modules/auth/application/guard/auth-team.guard';
 import { AuthType } from '@/modules/auth/domain/auth_type.enum';
-import { AdminRoleGuard } from '@/modules/authorization/infraestructure/policy/guard/admin-role.guard';
-import { AuthTeamGuard } from '@/modules/authorization/infraestructure/policy/guard/auth-team.guard';
 
 import { CreateEnviromentDto } from '../application/dto/create-enviroment.dto';
 import { EnviromentResponseDto } from '../application/dto/enviroment-response.dto';
